Apply formatting rules to UpdateInstaller

+++ Libraries/UpdateInstaller/SharpZipLib/Checksums/CRC32.cs
@@ -124,7 +124,7 @@ namespace ICSharpCode.SharpZipLib.Checksums
 
     internal static uint ComputeCrc32(uint oldCrc, byte value)
     {
-      return (uint)(Crc32.CrcTable[(oldCrc ^ value) & 0xFF] ^ (oldCrc >> 8));
+      return Crc32.CrcTable[(oldCrc ^ value) & 0xFF] ^ (oldCrc >> 8);
     }
 
     /// <summary>
@@ -139,7 +139,7 @@ namespace ICSharpCode.SharpZipLib.Checksums
     {
       get
       {
-        return (long)crc;
+        return crc;
       }
       set
       {
